Changing warning level to info for materialized collections
