+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer/MaterializedCollectionsUtils.cs
@@ -19,6 +19,8 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace MindTouchMaterializeCollectionsAnalyzer {
     internal class MaterializedCollectionsUtils {
@@ -26,6 +28,68 @@ namespace MindTouchMaterializeCollectionsAnalyzer {
         //--- Class Methods ---
         internal static bool IsCollection(ITypeSymbol symbol) {
             return (symbol != null && symbol.Interfaces.Any(x => x.Name == typeof(IEnumerable).Name || x.Name == typeof(IEnumerable<>).Name));
+        }
+
+        internal static bool IsAbstractCollectionType(TypeInfo typeInfo) {
+            if(typeInfo.Type == null || typeInfo.ConvertedType == null) {
+                return false;
+            }
+            return typeInfo.Type.IsAbstract && MaterializedCollectionsUtils.IsCollection(typeInfo.ConvertedType);
+        }
+
+        internal static bool ShouldReportOnCollectionNode(SemanticModel semanticModel, SyntaxNode argument) {
+            const bool returnIfUnknown = false;
+
+            // if the variable is an invocation syntax we can assume that it was returned materialized, if it's not an extension method
+            switch(argument.Kind()) {
+            case SyntaxKind.InvocationExpression:
+                var methodCallInfo = semanticModel.GetSymbolInfo(argument);
+                if(methodCallInfo.Symbol != null && methodCallInfo.Symbol.Kind == SymbolKind.Method) {
+                    var mSymbol = (IMethodSymbol)methodCallInfo.Symbol;
+
+                    // If the method is not an extension method, we assume it returned a materialized collection
+                    return mSymbol.IsExtensionMethod && mSymbol.ContainingNamespace.ToDisplayString().Equals("System.Linq");
+                }
+                break;
+            case SyntaxKind.IdentifierName:
+                var identifierInfo = semanticModel.GetSymbolInfo(argument);
+
+                // if this identifier came straight from a parameter, assume it is materialized
+                if(identifierInfo.Symbol != null && identifierInfo.Symbol.Kind == SymbolKind.Parameter) {
+
+                    //TODO(2016-02-24, yurig): if parameter is modified locally, we should warn
+                    return false;
+                }
+
+                // if this is a local identifier, look at where it is defined
+                if(identifierInfo.Symbol != null && identifierInfo.Symbol.Kind == SymbolKind.Local) {
+                    var declaration = identifierInfo.Symbol.DeclaringSyntaxReferences.FirstOrDefault();
+
+                    // if the declaration is an equals expression, dive into it.
+                    var equalsExpression = declaration?.GetSyntax()
+                        .ChildNodes()
+                        .FirstOrDefault(x => x.Kind() == SyntaxKind.EqualsValueClause);
+                    var firstChild = equalsExpression?.ChildNodes().FirstOrDefault();
+                    if(firstChild != null) {
+                        return ShouldReportOnCollectionNode(semanticModel, firstChild);
+                    }
+
+                    // if the variable was assigned somewhere else, find it
+                    var containingClass = declaration?.GetSyntax().FirstAncestorOrSelf<MethodDeclarationSyntax>();
+                    var localAssignment = containingClass?.DescendantNodes().OfType<AssignmentExpressionSyntax>()
+                        .Where(x => x.Left.IsKind(SyntaxKind.IdentifierName))
+                        .FirstOrDefault(x => (x.Left as IdentifierNameSyntax).Identifier.Text.Equals(((IdentifierNameSyntax)argument).Identifier.Text));
+                    if(localAssignment != null) {
+                        return ShouldReportOnCollectionNode(semanticModel, localAssignment.Right);
+                    }
+                }
+                break;
+            case SyntaxKind.SimpleMemberAccessExpression:
+
+                // Assume that member accesses are returned materialized
+                return false;
+            }
+            return returnIfUnknown;
         }
     }
 }
