+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer/Resources.resx
@@ -122,7 +122,7 @@
     <comment>An optional longer localizable description of the diagnostic.</comment>
   </data>
   <data name="AnalyzerMessageFormat" xml:space="preserve">
-    <value>Collection should be materialized to a specific type</value>
+    <value>Collection may or may not be materialized</value>
     <comment>The format-able message the diagnostic displays.</comment>
   </data>
   <data name="AnalyzerTitle" xml:space="preserve">
